merge-MainMenu-into-Map
merge and name new page Game, move map as component
