+++ client/src/components/Chat/Chat.scss
@@ -1,5 +1,4 @@
 .chat {
-    display: flex;
     flex-direction: column;
     border: 5px solid #3f91d4;
