+++ client/src/components/Map/Map.scss
@@ -1,0 +1,7 @@
+
+.map {
+    width: 640px;
+    height: 400px;
+    border: 5px solid black;
+    background-color: green;
+}
