+++ client/src/components/Map/Map.tsx
@@ -1,0 +1,13 @@
+import React from 'react';
+
+import './Map.scss';
+
+const Map: React.FC = () => {
+
+    return (
+    <div className='map'>
+        типа карта
+    </div>)
+}
+
+export default Map;
