+++ client/src/pages/Game/Game.scss
@@ -3,7 +3,12 @@
     border: 5px solid black;
 }
 
-.moveButton {
+.game-wrapper {
+    display: flex;
+    flex-direction: row;
+}
+
+.move-button {
     min-width: 50px;
     min-height: 50px;
 }
