--- client/src/pages/MainMenu/MainMenu.scss
@@ -1,7 +1,0 @@
-
-.user-resources {
-    display: inline-flex;
-    flex-direction: row;
-    flex-wrap: nowrap;
-    justify-content: space-evenly;
-}
