--- client/src/services/canvas/index.ts
@@ -1,4 +1,0 @@
-import useCanvas from "./useCanvas";
-import Canvas from "./Canvas";
-
-export { useCanvas, type Canvas };
